Removing unnecessary service
